fix(frontend): unify Heartbeat settings into the Priorities page
Move the proactive-check-in controls (opt-in, frequency, max daily)
out of the Settings tab and into the Priorities page above the
priorities markdown editor. Drop the standalone Heartbeat tab from
Settings so a new user is not confronted with "heartbeat" jargon as
a top-level setting.
