+++ frontend/src/extensions/settings.tsx
@@ -14,5 +14,5 @@ export function renderPremiumSettingsTab(_key: string, _isAdmin: boolean): React
 }
 
 export function showOssSettingsTabs(_isPremium: boolean, _isAdmin: boolean): string[] {
-  return ['model', 'heartbeat', 'telegram', 'channels', 'privacy'];
+  return ['model', 'telegram', 'channels', 'privacy'];
 }
