+++ frontend/src/pages/SettingsPage.tsx
@@ -5,11 +5,10 @@ import Input from '@/components/ui/input';
 import Button from '@/components/ui/button';
 import Select from '@/components/ui/select';
 import { Tabs, Tab } from '@heroui/tabs';
-import Checkbox from '@/components/ui/checkbox';
 import Field from '@/components/ui/field';
 import api from '@/api';
 import { toast } from '@/lib/toast';
-import { useModelConfig, useUpdateModelConfig, useUpdateProfile, useChannelConfig, useUpdateChannelConfig } from '@/hooks/queries';
+import { useModelConfig, useUpdateModelConfig, useChannelConfig, useUpdateChannelConfig } from '@/hooks/queries';
 import DataSharingConsentSection from '@/components/DataSharingConsentSection';
 import ChannelsPage from '@/pages/ChannelsPage';
 import type { AppShellContext } from '@/layouts/AppShell';
@@ -22,7 +21,7 @@ import {
 export default function SettingsPage() {
   const { tab } = useParams<{ tab: string }>();
   const navigate = useNavigate();
-  const { profile, reloadProfile, isPremium, isAdmin } = useOutletContext<AppShellContext>();
+  const { reloadProfile, isPremium, isAdmin } = useOutletContext<AppShellContext>();
 
   // Refresh profile whenever the settings page is opened.
   useEffect(() => {
@@ -39,7 +38,6 @@ export default function SettingsPage() {
   const visibleOssKeys = showOssSettingsTabs(isPremium, isAdmin);
   const ossTabs = [
     { key: 'model', label: 'Model' },
-    { key: 'heartbeat', label: 'Heartbeat' },
     { key: 'telegram', label: 'Telegram' },
     { key: 'channels', label: 'Channels' },
     { key: 'privacy', label: 'Privacy' },
@@ -55,7 +53,6 @@ export default function SettingsPage() {
     if (premiumContent) return premiumContent;
     switch (activeTab) {
       case 'model': return <ModelTab />;
-      case 'heartbeat': return profile ? <HeartbeatTab profile={profile} /> : null;
       case 'telegram': return <TelegramTab />;
       case 'channels': return <ChannelsPage />;
       case 'privacy': return <PrivacyTab />;
@@ -360,128 +357,6 @@ function ModelTab() {
         </Button>
       </div>
     </div>
-  );
-}
-
-// --- Storage Tab ---
-
-// --- Heartbeat Tab ---
-
-const HEARTBEAT_PRESETS = [
-  { value: '15m', label: 'Every 15 minutes' },
-  { value: '30m', label: 'Every 30 minutes' },
-  { value: '1h', label: 'Every hour' },
-  { value: '2h', label: 'Every 2 hours' },
-  { value: '4h', label: 'Every 4 hours' },
-  { value: '8h', label: 'Every 8 hours' },
-  { value: 'daily', label: 'Daily' },
-  { value: 'weekdays', label: 'Weekdays only' },
-  { value: 'weekly', label: 'Weekly' },
-] as const;
-
-function HeartbeatTab({
-  profile,
-}: {
-  profile: { heartbeat_opt_in: boolean; heartbeat_frequency: string; heartbeat_max_daily: number };
-}) {
-  const isPreset = HEARTBEAT_PRESETS.some((p) => p.value === profile.heartbeat_frequency);
-  const [form, setForm] = useState({
-    heartbeat_opt_in: profile.heartbeat_opt_in,
-    heartbeat_frequency: isPreset ? profile.heartbeat_frequency : 'custom',
-    custom_frequency: isPreset ? '' : profile.heartbeat_frequency,
-    heartbeat_max_daily: profile.heartbeat_max_daily,
-  });
-  const updateProfile = useUpdateProfile();
-
-  useEffect(() => {
-    const preset = HEARTBEAT_PRESETS.some((p) => p.value === profile.heartbeat_frequency);
-    setForm({
-      heartbeat_opt_in: profile.heartbeat_opt_in,
-      heartbeat_frequency: preset ? profile.heartbeat_frequency : 'custom',
-      custom_frequency: preset ? '' : profile.heartbeat_frequency,
-      heartbeat_max_daily: profile.heartbeat_max_daily,
-    });
-  }, [profile.heartbeat_opt_in, profile.heartbeat_frequency, profile.heartbeat_max_daily]);
-
-  const effectiveFrequency = form.heartbeat_frequency === 'custom'
-    ? form.custom_frequency
-    : form.heartbeat_frequency;
-
-  const handleSave = () => {
-    updateProfile.mutate(
-      {
-        heartbeat_opt_in: form.heartbeat_opt_in,
-        heartbeat_frequency: effectiveFrequency,
-        heartbeat_max_daily: form.heartbeat_max_daily,
-      },
-      {
-        onSuccess: () => toast.success('Heartbeat settings updated'),
-        onError: (e) => toast.error(e.message),
-      },
-    );
-  };
-
-  return (
-    <Card>
-      <div className="grid gap-4">
-        <div className="flex items-center gap-3">
-          <Checkbox
-            id="heartbeat-opt-in"
-            checked={form.heartbeat_opt_in}
-            onChange={(e) => setForm((prev) => ({ ...prev, heartbeat_opt_in: e.target.checked }))}
-          />
-          <label htmlFor="heartbeat-opt-in" className="text-sm">
-            Enable heartbeat check-ins
-          </label>
-        </div>
-        <p className="text-xs text-muted-foreground">
-          When enabled, your assistant will proactively send you reminders and updates based on your heartbeat items.
-        </p>
-        <Field label="Frequency">
-          <Select
-            value={form.heartbeat_frequency}
-            onChange={(e) => setForm((prev) => ({ ...prev, heartbeat_frequency: e.target.value }))}
-            disabled={!form.heartbeat_opt_in}
-          >
-            {HEARTBEAT_PRESETS.map((p) => (
-              <option key={p.value} value={p.value}>{p.label}</option>
-            ))}
-            <option value="custom">Custom interval</option>
-          </Select>
-        </Field>
-        {form.heartbeat_frequency === 'custom' && (
-          <Field label="Custom Interval">
-            <Input
-              value={form.custom_frequency}
-              onChange={(e) => setForm((prev) => ({ ...prev, custom_frequency: e.target.value }))}
-              disabled={!form.heartbeat_opt_in}
-              placeholder="e.g. 45m, 3h, 2d"
-            />
-            <p className="text-xs text-muted-foreground mt-1">
-              Use a number followed by m (minutes), h (hours), or d (days).
-            </p>
-          </Field>
-        )}
-        <Field label="Max daily check-ins">
-          <Input
-            type="number"
-            min={0}
-            value={form.heartbeat_max_daily}
-            onChange={(e) => setForm((prev) => ({ ...prev, heartbeat_max_daily: parseInt(e.target.value, 10) || 0 }))}
-            disabled={!form.heartbeat_opt_in}
-            placeholder="0"
-          />
-          <p className="text-xs text-muted-foreground mt-1">
-            Maximum number of heartbeat messages per day. Set to 0 to use the server default.
-          </p>
-        </Field>
-        <div className="flex justify-end">
-          <Button onClick={handleSave} disabled={updateProfile.isPending} isLoading={updateProfile.isPending}>
-            Save Heartbeat Settings
-          </Button>
-        </div>
-      </div>
-    </Card>
   );
 }
